# contract/openapi.yaml
openapi: 3.0.3
info:
  title: AI Lifestyle App API
  description: |
    API for the AI Lifestyle App - a comprehensive lifestyle management platform
    covering nutrition, fitness, wellness, and daily routines.
  version: 1.0.0
  contact:
    name: AI Lifestyle Team
    email: support@ailifestyle.app
servers:
  - url: https://api.ailifestyle.app/v1
    description: Production server
  - url: https://staging-api.ailifestyle.app/v1
    description: Staging server
  - url: http://localhost:4000/v1
    description: Local development
security:
  - BearerAuth: []
paths:
...
  /users/profile:
    get:
      summary: Get current user profile
      description: Returns the authenticated user's profile information
      operationId: getUserProfile
      tags:
        - Users
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      summary: Update user profile
      description: Updates the authenticated user's profile information
      operationId: updateUserProfile
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          description: Invalid profile data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # 2FA Endpoints
...
  /users/{userId}:
    get:
      summary: Get user by ID
      description: Retrieves basic user information by user ID
      operationId: getUserById
      tags:
        - Users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserBasicInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /encryption/user/{userId}:
    get:
      summary: Get user encryption info
      description: Retrieves user's public encryption key
      operationId: getUserEncryption
      tags:
        - Encryption
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Encryption info retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserEncryptionInfo'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found or encryption not set up
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from Cognito
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error type/code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        request_id:
          type: string
          description: Unique request identifier for debugging
        timestamp:
          type: string
          format: date-time
    ValidationErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
        - validation_errors
      properties:
        error:
          type: string
          enum: [VALIDATION_ERROR]
        message:
          type: string
          default: "Validation failed"
        validation_errors:
          type: array
          items:
            type: object
            required:
              - field
              - message
            properties:
              field:
                type: string
                description: Field that failed validation
              message:
                type: string
                description: Validation error message
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    UserProfile:
      type: object
      required:
        - userId
        - email
        - firstName
        - lastName
        - createdAt
        - updatedAt
        - emailVerified
      properties:
        userId:
          type: string
          format: uuid
          description: Unique user identifier
        email:
          type: string
          format: email
          description: User's email address
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        emailVerified:
          type: boolean
          description: Whether email is verified
        mfaEnabled:
          type: boolean
          description: Whether 2FA is enabled
          default: false
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
          description: User's phone number (E.164 format)
        dateOfBirth:
          type: string
          format: date
          description: User's date of birth
        timezone:
          type: string
          description: User's timezone (IANA format)
          example: "America/New_York"
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z\s-]+$'
        lastName:
          type: string
          minLength: 1
          maxLength: 50
          pattern: '^[a-zA-Z\s-]+$'
        phoneNumber:
          type: string
          pattern: '^\+?[1-9]\d{1,14}$'
        dateOfBirth:
          type: string
          format: date
        timezone:
          type: string
        preferences:
          $ref: '#/components/schemas/UserPreferences'
    UserPreferences:
      type: object
      properties:
        units:
          type: string
          enum: [metric, imperial]
          default: metric
          description: Measurement unit preference
        language:
          type: string
          pattern: '^[a-z]{2}(-[A-Z]{2})?$'
          default: en-US
          description: Language preference (ISO 639-1)
        notifications:
          type: object
          properties:
            email:
              type: boolean
              default: true
            push:
              type: boolean
              default: true
            sms:
              type: boolean
              default: false
        dietaryRestrictions:
          type: array
          items:
            type: string
            enum: [vegan, vegetarian, gluten-free, dairy-free, nut-free, halal, kosher]
        fitnessGoals:
          type: array
          items:
            type: string
            enum: [weight-loss, muscle-gain, endurance, flexibility, general-health]

    # Goal Management Schemas
...
    UserBasicInfo:
      type: object
      required:
        - userId
        - email
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        username:
          type: string
    UserEncryptionInfo:
      type: object
      required:
        - userId
        - publicKey
        - publicKeyId
      properties:
        userId:
          type: string
          format: uuid
        publicKey:
          type: string
          description: Base64 encoded public key
        publicKeyId:
          type: string
        salt:
          type: string
          description: Base64 encoded salt
        encryptedPrivateKey:
          type: string
          description: Base64 encoded encrypted private key
...